+++ new/index.php
@@ -1,21 +1,19 @@
 <?php
 $strategies = array();
+$strategy = $_GET['strategy']
 
-if (is_null($_GET['strategy'])) {
-    // No strategy specified
-    $response = array('response' => false, 'reason' => "Strategy not specified");
-    echo json_encode($response);
-    exit;
+if (is_null($strategy) {
+    responseError('Strategy not specified')
 }
-
-$strategy = $_GET[strategy];
 if (is_null($strategies[$strategy])) {
-    // Strategy not in array of defined strategies
-    $response = array('response' => false, 'reason' => "Unknown strategy");
-    echo json_encode($response);
-    exit;
+    responseError('Unknown Strategy')
 }
 
 // TODO: Implement game creation
-$response = array('response' => false, 'reason' => 'Not implemented yet!');
-echo json_encode($response);
+responseError('Not implemented yet!')
+
+function responseError($message) {
+    $response = array('response' => false, 'reason' => $message);
+    echo json_encode($response);
+    exit;
+}