+++ play/index.php
@@ -1,25 +1,26 @@
 <?php
-$width = 7
+$width = 7;
 
-if(is_null($_GET['pid'])) {
-    $response = array('response' => false, 'reason' => "PID not specified");
-    echo json_encode($response);
-    exit;
-}
 $pid = $_GET['pid']
-
-if(is_null($_GET['move'])) {
-    $response = array('response' => false, 'reason' => "Move not specified");
-    echo json_encode($response);
-    exit;
-}
 $move = $_GET['move']
 
+if(is_null($pid)) {
+    responseError('PID not specified');
+}
+if(is_null($move)) {
+    responseError('Move not specified');
+}
+if(!is_numeric($move)) {
+    responseError('Non-numeric move');
+}
+// No games are created yet, so all PIDs are unknown
+responseError('Unknown PID');
 if($move < 0 || $move >= $width) {
-    $response = array('response' => false, 'reason' => "Invalid slot, " . $move);
+    responseError("Invalid move $move");
+}
+
+function responseError($message) {
+    $response = array('response' => false, 'reason' => $message);
     echo json_encode($response);
     exit;
 }
-
-$response = array('response' => false, 'reason' => "Unknown PID");
-echo json_encode($response);